add: Fix panic when history command returns nothing
If you disable bash history files alltogether, "history 1" will not contain
anything on first launch of a terminal and therefore the regex
has nothing to parse, so we get an out of bounds array access.
So fix this.

Though it's a dirty silent fail now, I don't want error messages/warnings
everytime I open a shell.